Attempt to upsert and destroy stacks
PRs should use unique stacks for previews,
otherwise they'll stomp on each other.

This generates a unique stack name based on the PR
and upserts that with the Pulumi action if needed.

We need the generated Pulumi.$stack.yaml for the destroy operation,
but the two job will be separate workflow runs with no shared context.
Try to work around it by uploading the stack artifact,
and using a third-party GitHub action to download it
as the official download-artifact action doesn't support this.